Add files via upload
The basic use of the dictionary class on .NET Core in C# language is intended in the phoneBook application.
Application description:
The number of people to be registered will be entered from the keyboard.
Name and number will be entered with one space between them.
When the registration process is completed, the name of the person whose information is requested will be entered. If an unregistered name is entered, "Not found" will be printed.